+++ zedd-app/src/renderer.ts
@@ -49,22 +49,6 @@ const d = (...x: any[]) => console.log('renderer.ts', ...x)
 
 const isWin = process.platform === 'win32'
 
-// class Todo {
-//   name: string
-// }
-
-// createModelSchema(Todo, {
-//   title: optional(primitive()),
-//   user: optional(
-//     custom(
-//       (value) => value.name,
-//       () => SKIP,
-//     ),
-//   ),
-// })
-
-// serialize(new Todo()) // {}
-
 function showNotification(title: string, text: string, cb: () => void) {
   const notification = new Notification(title, {
     body: text,
@@ -137,8 +121,6 @@ async function setup() {
     platformState.setIntegrationMap()
   })
 
-  // await sleep(5000);
-  // importAndSaveClarityTasks();
   try {
     await platformState.loadStateFromFile()
   } catch (e) {
@@ -193,29 +175,6 @@ async function setup() {
           state.changingSliceTask = newSlice
         }
       },
-      // code for interactive notification. Disabled because it only works with a native module
-      // which isn't worth the hassle.
-      //   [state.currentTask.name.substring(0), formatInterval(when) + ' ' + state.currentTask.name],
-      //   ['Other...', formatInterval(when) + ' ' + '$$$OTHER$$$'],
-      //   (_, wargs) => {
-      //     const [start, end, taskName] = TimeSlice.parse(wargs.arguments)
-      //     const newSlice = new TimeSlice(
-      //       start,
-      //       end,
-      //       '$$$OTHER$$$' === taskName ? state.getUndefinedTask() : state.getTaskForName(taskName),
-      //     )
-      //     state.addSlice(newSlice)
-      //     if ('$$$OTHER$$$' === taskName) {
-      //       if (!currentWindow.isVisible()) {
-      //         currentWindow.show()
-      //       }
-      //       if (state.hoverMode) {
-      //         state.hoverMode = false
-      //       }
-      //       currentWindow.focus()
-      //       state.changingSliceTask = newSlice
-      //     }
-      //   },
     )
   }
 
@@ -272,6 +231,7 @@ async function setup() {
     }
     platformState.chromeExe = state.config.chromePath
     platformState.chromedriverExe = chromeDriverPath
+    await platformState.setIntegrationMap()
     return { chromeVersion, chromeDriverVersion: requiredChromeDriverVersion }
   }
   checkChromePath().catch((error) => state.addMessage(error.message))
@@ -319,17 +279,18 @@ async function setup() {
   const restoreUnmaximizedBoundsIfNotHoverMode = () =>
     !state.hoverMode && setBoundsSafe(currentWindow, state.bounds.normal)
 
-  const saveWindowBounds = ({ sender }: { sender: BrowserWindow }) => {
+  const saveWindowBounds = () => {
+    const win = getCurrentWindow()
     if (state && !state.hoverMode) {
-      if (sender.isMaximized()) {
+      if (win.isMaximized()) {
         state.bounds.maximized = true
       } else {
         state.bounds.maximized = false
-        state.bounds.normal = sender.getBounds()
+        state.bounds.normal = win.getBounds()
       }
     }
     if (state && state.hoverMode) {
-      state.bounds.hover = sender.getBounds()
+      state.bounds.hover = win.getBounds()
     }
   }
 
@@ -401,13 +362,11 @@ async function setup() {
 
         ...(0 === state.links.length ? [] : [{ type: 'separator' } as MenuItemConstructorOptions]),
 
-        ...state.links.map(
-          ([key, link]): MenuItemConstructorOptions => ({
-            label: 'Open in Browser: ' + key,
-            type: 'normal',
-            click: () => shell.openExternal(link),
-          }),
-        ),
+        ...state.links.map(([key, link]): MenuItemConstructorOptions => ({
+          label: 'Open in Browser: ' + key,
+          type: 'normal',
+          click: () => shell.openExternal(link),
+        })),
 
         { type: 'separator' },
 
@@ -459,10 +418,7 @@ async function setup() {
   const cleanupHoverModeAutorun = autorun(() => {
     currentWindow.setSkipTaskbar(state.hoverMode)
     currentWindow.setAlwaysOnTop(state.hoverMode)
-    // currentWindow.resizable = !state.hoverMode
     console.log('currentWindow.resizable', currentWindow.resizable)
-    // console.log('showing:', state.hoverMode, !currentWindow.isVisible)
-    // state.hoverMode && !currentWindow.isVisible && currentWindow.show()
     if (state.hoverMode) {
       const vertical = 'vertical' === state.config.keepHovering
       if (vertical) {