+++ zedd-platform/test/ott-import.ts
@@ -1,5 +1,7 @@
-import { PlatformIntegration } from "../src/platform-integration";
-import { OTTIntegration } from "../src";
+import { PlatformIntegration } from '../src/platform-integration'
+import { OTTIntegration } from '../src'
 
-const platformIntegration: PlatformIntegration = new OTTIntegration('', {headless: false, executablePath: '' })
-
+const platformIntegration: PlatformIntegration = new OTTIntegration('', {
+  headless: false,
+  executablePath: '',
+})