+++ zedd-win32/index.d.ts
@@ -1,7 +1,0 @@
-export function getMillisSinceLastUserInput(): number;
-export function getWindowInfos(): {
-  name: string;
-  foreground: boolean;
-  placement: "maximized" | "minimized" | "normal";
-  processName: string;
-};